backend/simulations/requirements.txt
diagrams==0.23.3
graphviz==0.20.1
Jinja2==3.1.2
MarkupSafe==2.1.2
typed-ast==1.5.4
googlemaps==4.10.0
python-dotenv==1.0.0
requests==2.28.2
python-dateutil==2.8.2
pytest==7.2.2
pytest-cov==4.0.0
uuid==1.30
pandas==1.5.3
google-cloud-bigquery==3.7.0
gunicorn==20.1.0
fastapi==0.95.0
uvicorn==0.21.1
httpx==0.23.3
pytest-dotenv==0.5.2